+++ config/reports/index.js
@@ -5,6 +5,8 @@ import OrganizationsDetailReport from './organizations-detail-report'
 import OrganizationsPlatformReport from './organizations-platform-report'
 import AdministrativeUnitsReport from "./administrative-units-report";
 import ChangeEventsReport from "./change-events-report";
+import OrganizationsNoAddressReport from "./organizations-no-address-report";
+import OrganizationsInFrontendReport from "./organizations-in-frontend-report";
 
 export default [
   OrganizationsReport,
@@ -14,4 +16,6 @@ export default [
   OrganizationsPlatformReport,
   AdministrativeUnitsReport,
   ChangeEventsReport,
+  OrganizationsNoAddressReport,
+  OrganizationsInFrontendReport,
 ];