+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 ## Unreleased
+- Query to check missing address URI's [OP-3784]
+- Create new query 'organisations shown in OP frontend' in Dashboard [OP-3774]
 - Moved AGB BRUGGE to correct graph [OP-3790]
 
 ### Deploy notes
 ```
+drc restart report-generation 
 drc restart migrations
 ```